#381 Turn off the check for a stripe clubhouse pro subscription when admins log in

+++ app/Http/Controllers/Auth/LoginController.php
@@ -8,6 +8,7 @@ use App\RoleUser;
 use App\Providers\StripeServiceProvider;
 use Illuminate\Foundation\Auth\AuthenticatesUsers;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
 
 class LoginController extends Controller
 {
@@ -43,14 +44,15 @@ class LoginController extends Controller
 
     protected function authenticated(Request $request, User $user)
     {
-        // TODO Check membership status
-        $stripe_user = StripeServiceProvider::getCustomer($user);
-        if (!is_null($stripe_user)) {
-            if ($stripe_user->delinquent || $stripe_user->subscriptions->total_count < 1) {
-                // Remove clubhouse role from user
-                $role = RoleUser::where(array(array('role_code', 'clubhouse'), array('user_id', $user->id)))->first();
-                if ($role) {
-                    $role->delete();
+        if (!\Gate::allows('view-admin-dashboard')) {
+            $stripe_user = StripeServiceProvider::getCustomer($user);
+            if (!is_null($stripe_user)) {
+                if ($stripe_user->delinquent || $stripe_user->subscriptions->total_count < 1) {
+                    // Remove clubhouse role from user
+                    $role = RoleUser::where(array(array('role_code', 'clubhouse'), array('user_id', $user->id)))->first();
+                    if ($role) {
+                        $role->delete();
+                    }
                 }
             }
         }
